requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.0=py36hf537a9a_0
asn1crypto=0.22.0=py36hb705621_1
bcolz=1.1.2=py36he76a132_0
bleach=2.0.0=py36h8fcea71_0
bokeh=0.12.10=py36hfd5be35_0
bzip2=1.0.6=h649919c_2
ca-certificates=2017.08.26=ha1e5d58_0
certifi=2017.11.5=py36ha569be9_0
cffi=1.10.0=py36h880867e_1
chardet=3.0.4=py36h96c241c_1
click=6.7=py36hec950be_0
cloudpickle=0.5.2=py36h379c7bd_0
cryptography=2.0.3=py36h22d4226_1
cycler=0.10.0=py36hfc81398_0
cymem=1.31.2=py36_0
cytoolz=0.8.2=py36h290905f_0
dask=0.16.0=py36h40f4302_0
dask-core=0.16.0=py36haa8de34_0
dbus=1.10.22=h50d9ad6_0
decorator=4.1.2=py36h69a1b52_0
dill=0.2.7.1=py36hdf89ffe_0
distributed=1.20.0=py36h1c9538a_0
django=1.11.3=py36_0
entrypoints=0.2.3=py36hd81d71f_2
expat=2.2.5=hb8e80ba_0
fontconfig=2.12.4=hffb9db1_2
freetype=2.8=h12048fb_1
ftfy=4.4.3=py36_0
gettext=0.19.8.1=h15daf44_3
glib=2.53.6=h33f6a65_2
hdf5=1.10.1=ha036c08_1
heapdict=1.0.0=py36h27a9ac6_0
html5lib=0.999999999=py36h79312fd_0
icu=58.2=h4b95b61_1
idna=2.6=py36h8628d0a_1
intel-openmp=2018.0.0=h8158457_8
ipykernel=4.6.1=py36h3208c25_0
ipython=6.2.1=py36h3dda519_1
ipython_genutils=0.2.0=py36h241746c_0
ipywidgets=7.0.5=py36h5142716_0
jbig=2.1=h4d881f8_0
jedi=0.10.2=py36h6325097_0
jinja2=2.9.6=py36hde4beb4_1
jpeg=9b=he5867d9_2
jsonschema=2.6.0=py36hb385e00_0
jupyter=1.0.0=py36h598a6cc_0
jupyter_client=5.1.0=py36hf6c435f_0
jupyter_console=5.2.0=py36hccf5b1c_1
jupyter_core=4.4.0=py36h79cf704_0
libcxx=4.0.1=h579ed51_0
libcxxabi=4.0.1=hebd6815_0
libedit=3.1=hb4e282d_0
libffi=3.2.1=h475c297_4
libgcc=4.8.5=hdbeacc1_10
libgfortran=3.0.1=h93005f0_2
libiconv=1.15=hdd342a3_7
libpng=1.6.32=hd1e8b91_4
libsodium=1.0.15=hd9e47c5_0
libtiff=4.0.9=h0dac147_0
libxml2=2.9.4=hf05c021_6
locket=0.2.0=py36hca03003_1
lzo=2.10=h362108e_2
markupsafe=1.0=py36h3a1e703_1
matplotlib=2.1.0=py36h5068139_0
mistune=0.8.1=py36h638d0ca_0
mkl=2018.0.1=hfbd8650_4
msgpack-python=0.4.8=py36h46767b2_0
murmurhash=0.26.4=py36_0
nbconvert=5.3.1=py36h810822e_0
nbformat=4.4.0=py36h827af21_0
ncurses=6.0=hd04f020_2
notebook=5.2.1=py36h640abe8_0
numexpr=2.6.2=py36h8fc668d_2
numpy=1.13.3=py36h2cdce51_0
olefile=0.44=py36ha08bf50_0
openssl=1.0.2m=h86d3e6a_1
pandas=0.21.0=py36hfed917e_1
pandoc=1.19.2.1=ha5e8f32_1
pandocfilters=1.4.2=py36h3b0b094_1
partd=0.3.8=py36hf5c4cb8_0
path.py=10.3.1=py36hd33c240_0
pathlib=1.0.1=py36_0
patsy=0.4.1=py36ha1b3fa5_0
pcre=8.41=hfb6ab37_1
pexpect=4.2.1=py36h3eac828_0
pickleshare=0.7.4=py36hf512f8e_0
pillow=4.2.1=py36h0263179_0
pip=9.0.1=py36h1555ced_4
plac=0.9.6=py36_0
preshed=1.0.0=py36_0
prompt_toolkit=1.0.15=py36haeda067_0
psutil=5.4.0=py36ha052210_0
ptyprocess=0.5.2=py36he6521c3_0
pycparser=2.18=py36h724b2fc_1
pygments=2.2.0=py36h240cd3f_0
pyopenssl=17.2.0=py36h5d7bf08_0
pyparsing=2.2.0=py36hb281f35_0
pyqt=5.6.0=py36he5c6137_6
pysocks=1.6.7=py36hfa33cec_1
pytables=3.4.2=py36hfbd7ab0_2
python=3.6.3=h5ce8c04_4
python-dateutil=2.6.1=py36h86d2abb_1
pytorch=0.2.0=py36_4cu75
pytz=2017.3=py36hf0bf824_0
pyyaml=3.12=py36h2ba1e63_1
pyzmq=16.0.3=py36he48b5ad_0
qt=5.6.2=h9975529_14
qtconsole=4.3.1=py36hd96c0ff_0
readline=7.0=hc1231fa_4
regex=2017.11.09=py36hb672b4b_0
requests=2.18.4=py36h4516966_1
scipy=1.0.0=py36h1de22e9_0
seaborn=0.8.1=py36h595ecd9_0
setuptools=36.5.0=py36h2134326_0
simplegeneric=0.8.1=py36he5b5b09_0
sip=4.18.1=py36h2824476_2
six=1.11.0=py36h0e22d5e_1
sortedcontainers=1.5.7=py36ha982688_0
spacy=1.9.0=py36h4e87756_1
sqlite=3.20.1=h7e4c145_2
statsmodels=0.8.0=py36h9c68fc9_0
tblib=1.3.2=py36hda67792_0
termcolor=1.1.0=py36_0
terminado=0.6=py36h656782e_0
testpath=0.3.1=py36h625a49b_0
thinc=6.5.2=py36_0
tk=8.6.7=h35a86e2_3
toolz=0.8.2=py36h7b95164_0
torchvision=0.1.9=py36_1
tornado=4.5.2=py36h468dda9_0
tqdm=4.19.4=py36he502594_0
traitlets=4.3.2=py36h65bd3ce_0
ujson=1.35=py36_0
urllib3=1.22=py36h68b9469_0
wcwidth=0.1.7=py36h8c6ec74_0
webencodings=0.5.1=py36h3b9701d_1
wheel=0.30.0=py36h5eb2c71_1
widgetsnbextension=3.0.8=py36h34759f8_0
wrapt=1.10.11=py36hc29e774_0
xz=5.2.3=h0278029_2
yaml=0.1.7=hc338f04_2
zeromq=4.2.2=ha360ad0_2
zict=0.1.3=py36h71da714_0
zlib=1.2.11=hf3cbc9b_2
